fix(x): show the actual next meeting in the sidebar
